feat: service 토대 마련
