+++ src/main/java/com/dku/council/domain/ticket/service/TicketService.java
@@ -1,0 +1,21 @@
+package com.dku.council.domain.ticket.service;
+
+import com.dku.council.domain.ticket.model.dto.TicketDto;
+import com.dku.council.domain.ticket.model.dto.TicketEventDto;
+
+import java.util.List;
+
+public class TicketService {
+
+    public List<TicketEventDto> list() {
+        return null;
+    }
+
+    public TicketDto myTicket(Long userId, Long ticketEventId) {
+        return null;
+    }
+
+    public TicketDto enroll(Long userId) {
+        return null;
+    }
+}
