+++ src/main/java/com/dku/council/infra/nhn/exception/InvalidCaptchaException.java
@@ -1,0 +1,10 @@
+package com.dku.council.infra.nhn.exception;
+
+import com.dku.council.global.error.exception.LocalizedMessageException;
+import org.springframework.http.HttpStatus;
+
+public class InvalidCaptchaException extends LocalizedMessageException {
+    public InvalidCaptchaException() {
+        super(HttpStatus.BAD_REQUEST, "invalid.captcha");
+    }
+}
