+++ src/main/java/com/dku/council/infra/nhn/model/NhnMessage.java
@@ -1,4 +1,4 @@
-package com.dku.council.infra.nhn.model.dto;
+package com.dku.council.infra.nhn.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
