+++ src/main/java/com/dku/council/infra/nhn/service/CaptchaService.java
@@ -1,0 +1,16 @@
+package com.dku.council.infra.nhn.service;
+
+import com.dku.council.infra.nhn.exception.InvalidCaptchaException;
+import com.dku.council.infra.nhn.model.Captcha;
+
+// TODO Implement
+public class CaptchaService {
+
+    public Captcha requestCaptcha() {
+        return null;
+    }
+
+    public void verifyCaptcha(String captchaKey, String captchaValue) {
+        throw new InvalidCaptchaException();
+    }
+}
